restructure code

+++ src-cuda/cudaOperator.cu
@@ -1,5 +1,7 @@
-#include <iostream>
 #include "matrix.h"
+
+#include <iostream>
+
 #define INDEX(r, c, width) ((r) * (width) + (c))
 #define MBLK 16
 #define LBLK 32
@@ -12,81 +14,81 @@ static inline int updiv(int n, int d) {
     return (n + d - 1) / d;
 }
 
-__device__ static inline int sigmoid(double x) {
+__device__ static inline int sigmoid(float x) {
     return 1 / (1 + exp(x));
 }
 
-__device__ static inline double devTanh(double x) {
+__device__ static inline float devTanh(float x) {
     return tanh(x);
 }
 
 // kernel functions
-__global__ void cudaMatAddKernel(double* src1, double* src2, double* dst, int bound) {
-    int i = blockIdx.x * blockDim.x+ threadIdx.x;
-    if (i < bound){
+__global__ void cudaMatAddKernel(float *src1, float *src2, float *dst, int bound) {
+    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    if (i < bound) {
         dst[i] = src1[i] + src2[i];
     }
 }
 
-__global__ void cudaMatDotKernel(double* src1, double* src2, double* dst, int bound) {
-    int i = blockIdx.x * blockDim.x+ threadIdx.x;
-    if (i < bound){
+__global__ void cudaMatDotKernel(float *src1, float *src2, float *dst, int bound) {
+    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    if (i < bound) {
         dst[i] = src1[i] * src2[i];
     }
 }
 
-__global__ void cudaMatMulKernel(int M, int N, double* dmatA,double* dmatB, double * dmatC) {
-    int i = blockIdx.x * blockDim.x+ threadIdx.x;
-    int j = blockIdx.y * blockDim.y+ threadIdx.y;
-    if (i>= M || j >= N) {
+__global__ void cudaMatMulKernel(int M, int N, float *dmatA, float *dmatB, float *dmatC) {
+    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    int j = blockIdx.y * blockDim.y + threadIdx.y;
+    if (i >= M || j >= N) {
         return;
     }
     float sum = 0.0;
     for (int k = 0; k < N; k++) {
-        sum += dmatA[INDEX(i,k,N)] * dmatB[INDEX(k,j,N)];
+        sum += dmatA[INDEX(i, k, N)] * dmatB[INDEX(k, j, N)];
     }
-    dmatC[INDEX(i,j,N)] = sum;
+    dmatC[INDEX(i, j, N)] = sum;
 }
 
-__global__ void cudaSigmoidKernel(double* src, double* dst, int length) {
+__global__ void cudaSigmoidKernel(float *src, float *dst, int length) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i > length) return;
     dst[i] = sigmoid(src[i]);
 }
 
-__global__ void cudaTanhKernel(double* src, double* dst, int length) {
+__global__ void cudaTanhKernel(float *src, float *dst, int length) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i > length) return;
     dst[i] = devTanh(src[i]);
 }
 
 // host functions
-void CudaOperator::setup(int size, double* data) {
-    std::cout<<"cuda setup...\n";
-    cudaMemcpy(cuData, data, sizeof(double) * size, cudaMemcpyHostToDevice);
+void CudaOperator::setup(int size, float *data) {
+    std::cout << "cuda setup...\n";
+    cudaMemcpy(cuData, data, sizeof(float) * size, cudaMemcpyHostToDevice);
 }
 
-void CudaOperator::cuAdd(double* src1, double *src2, double* dst, int M, int N) {
-    std::cout<<"cuAdd()\n";
+void CudaOperator::cuAdd(float *src1, float *src2, float *dst, int M, int N) {
+    std::cout << "cuAdd()\n";
     int elements = M * N;
-    int size = elements * sizeof(double);
+    int size = elements * sizeof(float);
     // Allocate vectors in device memory
-    double* d_A;
+    float *d_A;
     cudaMalloc(&d_A, size);
-    double* d_B;
+    float *d_B;
     cudaMalloc(&d_B, size);
-    double* d_C;
+    float *d_C;
     cudaMalloc(&d_C, size);
-    
+
     // Copy vectors from host memory to device memory
     cudaMemcpy(d_A, src1, size, cudaMemcpyHostToDevice);
     cudaMemcpy(d_B, src2, size, cudaMemcpyHostToDevice);
-    
+
     // Invoke kernel
     int threadsPerBlock = MBLK * MBLK;
     int blocksPerGrid = (elements + threadsPerBlock - 1) / threadsPerBlock;
     cudaMatAddKernel<<<blocksPerGrid, threadsPerBlock>>>(d_A, d_B, d_C, elements);
-    
+
     // copy result
     cudaMemcpy(dst, d_C, size, cudaMemcpyDeviceToHost);
 
@@ -96,15 +98,15 @@ void CudaOperator::cuAdd(double* src1, double *src2, double* dst, int M, int N) 
     cudaFree(d_C);
 }
 
-void CudaOperator::cuMul(double* A, double* B, double* C, int M, int N) {
+void CudaOperator::cuMul(float *A, float *B, float *C, int M, int N) {
     int elements = M * N;
-    int size = elements * sizeof(double);
+    int size = elements * sizeof(float);
     // Allocate vectors in device memory
-    double* d_A;
+    float *d_A;
     cudaMalloc(&d_A, size);
-    double* d_B;
+    float *d_B;
     cudaMalloc(&d_B, size);
-    double* d_C;
+    float *d_C;
     cudaMalloc(&d_C, size);
 
     // Copy vectors from host memory to device memory
@@ -125,17 +127,17 @@ void CudaOperator::cuMul(double* A, double* B, double* C, int M, int N) {
     cudaFree(d_C);
 }
 
-void CudaOperator::cuDot(double* A, double* B, double* C, int M, int N) {
+void CudaOperator::cuDot(float *A, float *B, float *C, int M, int N) {
     int elements = M * N;
     int threadsPerBlock = MBLK * MBLK;
     int blocksPerGrid = updiv(elements, threadsPerBlock);
-    int size = elements * sizeof(double);
+    int size = elements * sizeof(float);
     // Allocate vectors in device memory
-    double* d_A;
+    float *d_A;
     cudaMalloc(&d_A, size);
-    double* d_B;
+    float *d_B;
     cudaMalloc(&d_B, size);
-    double* d_C;
+    float *d_C;
     cudaMalloc(&d_C, size);
 
     // Copy matrix from host memory to device memory
@@ -154,12 +156,12 @@ void CudaOperator::cuDot(double* A, double* B, double* C, int M, int N) {
     cudaFree(d_C);
 }
 
-void CudaOperator::cuSigmoid(double* src, double* dst, int length) {
-    int size = length * sizeof(double);
+void CudaOperator::cuSigmoid(float *src, float *dst, int length) {
+    int size = length * sizeof(float);
     // Allocate vectors in device memory
-    double* d_src;
+    float *d_src;
     cudaMalloc(&d_src, size);
-    double* d_dst;
+    float *d_dst;
     cudaMalloc(&d_dst, size);
 
     // Copy matrix from host memory to device memory
@@ -168,8 +170,8 @@ void CudaOperator::cuSigmoid(double* src, double* dst, int length) {
     // Invoke
     int threadsPerBlock = MBLK * MBLK;
     int blocksPerGrid = updiv(length, threadsPerBlock);
-    cudaSigmoidKernel<<<blocksPerGrid, threadsPerBlock>>>(d_src, d_dst, length); 
-    
+    cudaSigmoidKernel<<<blocksPerGrid, threadsPerBlock>>>(d_src, d_dst, length);
+
     // copy result
     cudaMemcpy(dst, d_dst, size, cudaMemcpyDeviceToHost);
 
@@ -178,12 +180,12 @@ void CudaOperator::cuSigmoid(double* src, double* dst, int length) {
     cudaFree(d_dst);
 }
 
-void CudaOperator::cuTanh(double* src, double* dst, int length) {
-    int size = length * sizeof(double);
+void CudaOperator::cuTanh(float *src, float *dst, int length) {
+    int size = length * sizeof(float);
     // Allocate vectors in device memory
-    double* d_src;
+    float *d_src;
     cudaMalloc(&d_src, size);
-    double* d_dst;
+    float *d_dst;
     cudaMalloc(&d_dst, size);
 
     // Copy matrix from host memory to device memory
@@ -191,10 +193,10 @@ void CudaOperator::cuTanh(double* src, double* dst, int length) {
 
     int threadsPerBlock = MBLK * MBLK;
     int blocksPerGrid = updiv(length, threadsPerBlock);
-    cudaSigmoidKernel<<<blocksPerGrid, threadsPerBlock>>>(d_src, d_dst, length);  
-    
+    cudaSigmoidKernel<<<blocksPerGrid, threadsPerBlock>>>(d_src, d_dst, length);
+
     // copy result
-    cudaMemcpy(dst, d_dst, size, cudaMemcpyDeviceToHost); 
+    cudaMemcpy(dst, d_dst, size, cudaMemcpyDeviceToHost);
 
     // Free device memory
     cudaFree(d_src);
