+++ src-cuda/include/matrix.h
@@ -5,29 +5,30 @@
 #ifndef RNN_DATA_H
 #define RNN_DATA_H
 
+#include "cudaOperator.h"
+#include "device.h"
+
 #include <random>
 #include <cassert>
 #include <cstring>
 #include <string>
 #include <iostream>
-#include "cudaOperator.h"
-#include "device.h"
 
 using namespace std;
 
 class Matrix {
 public:
-    Matrix() : _data(nullptr), _M(0), _N(0), _size(0), _dev(SEQ), _cu() { 
-        _dev = SEQ; 
+    Matrix() : _data(nullptr), _M(0), _N(0), _size(0), _dev(SEQ), _cu() {
+        _dev = SEQ;
     };
 
     Matrix(int M, int N, bool zero = true) : _M(M), _N(N), _size(M * N), _cu() {
         assert(M > 0);
         assert(N > 0);
-        _data = new double[_size]();
+        _data = new float[_size]();
         if (!zero) {
             for (int i = 0; i < _size; i++) {
-                _data[i] = static_cast<double>(_rd());
+                _data[i] = static_cast<float>(_rd());
             }
         }
     }
@@ -35,14 +36,14 @@ public:
     Matrix(int M, int N, Device device) : _M(M), _N(N), _size(M * N), _dev(device), _cu() {
         assert(M > 0);
         assert(N > 0);
-        _data = new double[_size]();
-        std::cout<<"constructor cuda\n";
+        _data = new float[_size]();
+        std::cout << "constructor cuda\n";
         if (device == GPU) {
             _cu = new CudaOperator();
         }
     }
 
-    Matrix(double *data, int M, int N) : _data(data), _M(M), _N(N), _size(M * N), _dev(SEQ), _cu() {
+    Matrix(float *data, int M, int N) : _data(data), _M(M), _N(N), _size(M * N), _dev(SEQ), _cu() {
         assert(data != nullptr);
         assert(M > 0);
         assert(N > 0);
@@ -52,8 +53,8 @@ public:
         _M = m._M;
         _N = m._N;
         _size = m._size;
-        _data = new double[m._size];
-        memcpy(_data, m._data, m._size * sizeof(double));
+        _data = new float[m._size];
+        memcpy(_data, m._data, m._size * sizeof(float));
     }
 
     ~Matrix() {
@@ -62,8 +63,11 @@ public:
     };
 
     Matrix operator-() const;
-    Matrix operator-(const double &) const;
-    Matrix operator+(const double &) const;
+
+    Matrix operator-(const float &) const;
+
+    Matrix operator+(const float &) const;
+
     Matrix &operator=(const Matrix &);
 
     int size();
@@ -78,17 +82,17 @@ public:
 
     Matrix dot(Matrix &);
 
-    double *data();
+    float *data();
 
     // CUDA
-    CudaOperator *_cu; 
+    CudaOperator *_cu;
     Device _dev;
 
 private:
     int _M;
     int _N;
     int _size;
-    double *_data;
+    float *_data;
     static mt19937_64 _rd;
 };
