+++ src-cuda/matrix.cpp
@@ -2,15 +2,16 @@
 // Created by yuwang on 2020-04-22.
 //
 
-#include <matrix.h>
+#include "matrix.h"
+
 #include <cstring>
 
 Matrix &Matrix::operator=(const Matrix &m) {
     if (&m == this) {
         return *this;
     }
-    this->_data = new double[m._size];
-    memcpy(this->_data, m._data, m._size*sizeof(double));
+    this->_data = new float[m._size];
+    memcpy(this->_data, m._data, m._size * sizeof(float));
 
     this->_M = m._M;
     this->_N = m._N;
@@ -39,7 +40,7 @@ Matrix Matrix::T() {
     if (_size == 0) {
         return Matrix();
     }
-    auto *data = new double[this->_size];
+    auto *data = new float[this->_size];
     auto M = _N;
     auto N = _M;
 
@@ -57,7 +58,7 @@ Matrix Matrix::add(Matrix &d) {
     assert(_N == d._N);
 
     // seq
-    auto *data = new double[_size];
+    auto *data = new float[_size];
     if (this->_dev == SEQ) {
         for (int i = 0; i < _M; i++) {
             for (int j = 0; j < _N; j++) {
@@ -78,7 +79,7 @@ Matrix Matrix::dot(Matrix &d) {
     assert(_M == d._M);
     assert(_N == d._N);
 
-    auto *data = new double[_size];
+    auto *data = new float[_size];
     if (this->_dev == SEQ) {
         for (int i = 0; i < _M; i++) {
             for (int j = 0; j < _N; j++) {
@@ -99,7 +100,7 @@ Matrix Matrix::mul(Matrix &d) {
     assert(_N == d._M);
 
     /* do calculation */
-    auto *data = new double[_M * d._N]();
+    auto *data = new float[_M * d._N]();
     if (this->_dev == SEQ) {
         for (int i = 0; i < _M; i++) {
             for (int j = 0; j < d._N; j++) {
@@ -118,28 +119,28 @@ Matrix Matrix::mul(Matrix &d) {
     return Matrix(data, _M, d._N);
 }
 
-double *Matrix::data() {
+float *Matrix::data() {
     return _data;
 }
 
 Matrix Matrix::operator-() const {
-    auto *data = new double[_size];
+    auto *data = new float[_size];
     for (int i = 0; i < _size; i++) {
         data[i] = -_data[i];
     }
     return Matrix(data, _M, _N);
 }
 
-Matrix Matrix::operator-(const double &num) const {
-    auto *data = new double[_size];
+Matrix Matrix::operator-(const float &num) const {
+    auto *data = new float[_size];
     for (int i = 0; i < _size; i++) {
         data[i] = _data[i] - num;
     }
     return Matrix(data, _M, _N);
 }
 
-Matrix Matrix::operator+(const double &num) const {
-    auto *data = new double[_size];
+Matrix Matrix::operator+(const float &num) const {
+    auto *data = new float[_size];
     for (int i = 0; i < _size; i++) {
         data[i] = data[i] + num;
     }
