+++ src-cuda/sigmoid.cpp
@@ -2,10 +2,10 @@
 // Created by yuwang4 on 2020-04-21.
 //
 
-#include <sigmoid.h>
+#include "sigmoid.h"
 
 Matrix Sigmoid::forward(Matrix m) {
-    auto *__data = new double[m.size()];
+    auto *__data = new float[m.size()];
     auto *_data = m.data();
     if (m._dev == SEQ) {
         for (int i = 0; i < m.size(); i++) {
@@ -16,7 +16,7 @@ Matrix Sigmoid::forward(Matrix m) {
     if (m._dev == GPU) {
         m._cu->cuSigmoid(_data, __data, m.size());
     }
-    
+
     _res = Matrix(__data, m.size(0), m.size(1));
     return _res;
 }
